data/requirements.txt
mysql-connector==2.2.9
sastrawi==1.0.1